+++ tasks.py
@@ -11,9 +11,12 @@ def start_development(c):
     r = c.run('sudo docker ps -f name=dev-haproxy', hide=True)
     if 'dev-haproxy' not in r.stdout:
         print('Starting HAproxy')
-        haproxycfg = os.path.join(os.path.abspath('.'), 'install/haproxy.cfg')
+        haproxycfg = os.path.join(os.path.abspath('.'), 'ansible/templates/haproxy.cfg.j2')
+        with open('/tmp/haproxy.cfg', 'w') as f:
+            content = open(haproxycfg, 'r').read().replace('{{ lpos_port.stdout }}', '8000')
+            f.write(content)
         c.run('sudo docker run --rm --name copier-haproxy -v dev-haproxy:/app -d alpine sleep 3')
-        c.run(f'sudo docker cp {haproxycfg} copier-haproxy:/app/')
+        c.run('sudo docker cp /tmp/haproxy.cfg copier-haproxy:/app/haproxy.cfg')
         cmd = [
             'sudo docker run --name dev-haproxy --rm --cap-add=NET_ADMIN',
             '--add-host=host.docker.internal:host-gateway --sysctl net.ipv4.ip_unprivileged_port_start=0',