+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * the first IP-Range, that is created for a VLAN-purpose, is saved as default for this purpose. If another IpPool for this VLAN-purpose is created, this default range is prefilled in the interface.
     * also a option on settings screen exists, to modify those default_ips
+  * added setting to overwrite SSO IP, to prevent system from doing a nslookup
   * optimized naming and sorting of Switches an IpPools on TablesScreen
   * smartified invoker for resetting switchtches